+++ src/graphics/mlx_wrappers/image_to_window.c
@@ -1,6 +1,6 @@
 #include "so_long.h"
 
-int32_t	image_to_window(mlx_t *mlx, mlx_image_t *image, size_t x, size_t y)
+int32_t	image_to_window(mlx_t *mlx, mlx_image_t *image, uint32_t x, uint32_t y)
 {
 	int32_t	return_value;
